requirements.txt
bcrypt==4.0.1
gunicorn
blinker==1.6.3
certifi==2016.2.28
cffi==1.16.0
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.3.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.0
idna==3.4
pyjwt==2.3.0
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==9.0.1
psycopg2-binary==2.9.9
pycparser==2.21
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
Werkzeug==2.3.0
WTForms==3.0.1
